+++ src/khoj/database/adapters/__init__.py
@@ -357,22 +357,52 @@ class ClientApplicationAdapters:
 
 class ConversationAdapters:
     @staticmethod
-    def get_conversation_by_user(user: KhojUser, client_application: ClientApplication = None):
-        conversation = Conversation.objects.filter(user=user, client=client_application)
+    def get_conversation_by_user(
+        user: KhojUser, client_application: ClientApplication = None, conversation_id: int = None, slug: str = None
+    ):
+        if conversation_id:
+            conversation = Conversation.objects.filter(user=user, client=client_application, id=conversation_id)
+        elif slug:
+            conversation = Conversation.objects.filter(user=user, client=client_application, slug=slug)
+        else:
+            conversation = Conversation.objects.filter(user=user, client=client_application)
         if conversation.exists():
             return conversation.first()
         return Conversation.objects.create(user=user, client=client_application)
 
     @staticmethod
-    async def aget_conversation_by_user(user: KhojUser, client_application: ClientApplication = None):
-        conversation = Conversation.objects.filter(user=user, client=client_application)
-        if await conversation.aexists():
-            return await conversation.afirst()
-        return await Conversation.objects.acreate(user=user, client=client_application)
+    def get_conversation_sessions(user: KhojUser, client_application: ClientApplication = None):
+        return Conversation.objects.filter(user=user, client=client_application)
 
     @staticmethod
-    async def adelete_conversation_by_user(user: KhojUser):
-        return await Conversation.objects.filter(user=user).adelete()
+    def get_conversation_by_id(conversation_id: int):
+        return Conversation.objects.filter(id=conversation_id).first()
+
+    @staticmethod
+    async def acreate_conversation_session(
+        user: KhojUser, client_application: ClientApplication = None, slug: str = None
+    ):
+        return await Conversation.objects.acreate(user=user, client=client_application, slug=slug)
+
+    @staticmethod
+    async def aget_conversation_by_user(
+        user: KhojUser, client_application: ClientApplication = None, conversation_id: int = None, slug: str = None
+    ):
+        if conversation_id:
+            conversation = Conversation.objects.filter(user=user, client=client_application, id=conversation_id)
+        else:
+            conversation = Conversation.objects.filter(user=user, client=client_application, slug=slug)
+        if await conversation.aexists():
+            return await conversation.afirst()
+        return await Conversation.objects.acreate(user=user, client=client_application, slug=slug)
+
+    @staticmethod
+    async def adelete_conversation_by_user(
+        user: KhojUser, client_application: ClientApplication = None, slug: str = None, conversation_id: int = None
+    ):
+        if conversation_id:
+            return await Conversation.objects.filter(user=user, client=client_application, id=conversation_id).adelete()
+        return await Conversation.objects.filter(user=user, client=client_application, slug=slug).adelete()
 
     @staticmethod
     def has_any_conversation_config(user: KhojUser):
@@ -433,12 +463,24 @@ class ConversationAdapters:
         return await ChatModelOptions.objects.filter().afirst()
 
     @staticmethod
-    def save_conversation(user: KhojUser, conversation_log: dict, client_application: ClientApplication = None):
-        conversation = Conversation.objects.filter(user=user, client=client_application)
-        if conversation.exists():
-            conversation.update(conversation_log=conversation_log)
+    def save_conversation(
+        user: KhojUser,
+        conversation_log: dict,
+        client_application: ClientApplication = None,
+        conversation_id: int = None,
+        user_message: str = None,
+    ):
+        slug = user_message.strip()[:200] if user_message else None
+        if conversation_id:
+            conversation = Conversation.objects.filter(user=user, client=client_application, id=conversation_id)
         else:
-            Conversation.objects.create(user=user, conversation_log=conversation_log, client=client_application)
+            conversation = Conversation.objects.filter(user=user, client=client_application)
+        if conversation.exists():
+            conversation.update(conversation_log=conversation_log, slug=slug)
+        else:
+            Conversation.objects.create(
+                user=user, conversation_log=conversation_log, client=client_application, slug=slug
+            )
 
     @staticmethod
     def get_conversation_processor_options():
