+++ src/khoj/routers/api.py
@@ -5,13 +5,12 @@ import math
 import os
 import time
 import uuid
-from typing import Any, Dict, List, Optional, Union
-from urllib.parse import unquote
+from typing import Any, List, Optional, Union
 
 from asgiref.sync import sync_to_async
 from fastapi import APIRouter, Depends, File, HTTPException, Request, UploadFile
 from fastapi.requests import Request
-from fastapi.responses import Response, StreamingResponse
+from fastapi.responses import Response
 from starlette.authentication import requires
 
 from khoj.configure import configure_server
@@ -25,19 +24,11 @@ from khoj.processor.conversation.offline.chat_model import extract_questions_off
 from khoj.processor.conversation.offline.whisper import transcribe_audio_offline
 from khoj.processor.conversation.openai.gpt import extract_questions
 from khoj.processor.conversation.openai.whisper import transcribe_audio
-from khoj.processor.conversation.prompts import help_message, no_entries_found
-from khoj.processor.conversation.utils import save_to_conversation_log
-from khoj.processor.tools.online_search import search_with_google
 from khoj.routers.helpers import (
     ApiUserRateLimiter,
     CommonQueryParams,
     ConversationCommandRateLimiter,
-    agenerate_chat_response,
-    get_conversation_command,
-    is_ready_to_chat,
-    text_to_image,
     update_telemetry_state,
-    validate_conversation_config,
 )
 from khoj.search_filter.date_filter import DateFilter
 from khoj.search_filter.file_filter import FileFilter
@@ -45,14 +36,7 @@ from khoj.search_filter.word_filter import WordFilter
 from khoj.search_type import image_search, text_search
 from khoj.utils import constants, state
 from khoj.utils.config import GPT4AllProcessorModel
-from khoj.utils.helpers import (
-    AsyncIteratorWrapper,
-    ConversationCommand,
-    command_descriptions,
-    get_device,
-    is_none_or_empty,
-    timer,
-)
+from khoj.utils.helpers import ConversationCommand, timer
 from khoj.utils.rawconfig import SearchResponse
 from khoj.utils.state import SearchType
 
@@ -222,81 +206,6 @@ def update(
     return {"status": "ok", "message": "khoj reloaded"}
 
 
-@api.get("/chat/starters", response_class=Response)
-@requires(["authenticated"])
-async def chat_starters(
-    request: Request,
-    common: CommonQueryParams,
-) -> Response:
-    user: KhojUser = request.user.object
-    starter_questions = await ConversationAdapters.aget_conversation_starters(user)
-    return Response(content=json.dumps(starter_questions), media_type="application/json", status_code=200)
-
-
-@api.get("/chat/history")
-@requires(["authenticated"])
-def chat_history(
-    request: Request,
-    common: CommonQueryParams,
-):
-    user = request.user.object
-    validate_conversation_config()
-
-    # Load Conversation History
-    meta_log = ConversationAdapters.get_conversation_by_user(
-        user=user, client_application=request.user.client_app
-    ).conversation_log
-
-    update_telemetry_state(
-        request=request,
-        telemetry_type="api",
-        api="chat",
-        **common.__dict__,
-    )
-
-    return {"status": "ok", "response": meta_log.get("chat", [])}
-
-
-@api.delete("/chat/history")
-@requires(["authenticated"])
-async def clear_chat_history(
-    request: Request,
-    common: CommonQueryParams,
-):
-    user = request.user.object
-
-    # Clear Conversation History
-    await ConversationAdapters.adelete_conversation_by_user(user)
-
-    update_telemetry_state(
-        request=request,
-        telemetry_type="api",
-        api="clear_chat_history",
-        **common.__dict__,
-    )
-
-    return {"status": "ok", "message": "Conversation history cleared"}
-
-
-@api.get("/chat/options", response_class=Response)
-@requires(["authenticated"])
-async def chat_options(
-    request: Request,
-    common: CommonQueryParams,
-) -> Response:
-    cmd_options = {}
-    for cmd in ConversationCommand:
-        cmd_options[cmd.value] = command_descriptions[cmd]
-
-    update_telemetry_state(
-        request=request,
-        telemetry_type="api",
-        api="chat_options",
-        **common.__dict__,
-    )
-    return Response(content=json.dumps(cmd_options), media_type="application/json", status_code=200)
-
-
 @api.post("/transcribe")
 @requires(["authenticated"])
 async def transcribe(
@@ -356,139 +265,6 @@ async def transcribe(
     # Return the spoken text
     content = json.dumps({"text": user_message})
     return Response(content=content, media_type="application/json", status_code=200)
-
-
-@api.get("/chat", response_class=Response)
-@requires(["authenticated"])
-async def chat(
-    request: Request,
-    common: CommonQueryParams,
-    q: str,
-    n: Optional[int] = 5,
-    d: Optional[float] = 0.18,
-    stream: Optional[bool] = False,
-    rate_limiter_per_minute=Depends(
-        ApiUserRateLimiter(requests=5, subscribed_requests=60, window=60, slug="chat_minute")
-    ),
-    rate_limiter_per_day=Depends(
-        ApiUserRateLimiter(requests=5, subscribed_requests=600, window=60 * 60 * 24, slug="chat_day")
-    ),
-) -> Response:
-    user: KhojUser = request.user.object
-    q = unquote(q)
-
-    await is_ready_to_chat(user)
-    conversation_command = get_conversation_command(query=q, any_references=True)
-
-    await conversation_command_rate_limiter.update_and_check_if_valid(request, conversation_command)
-
-    q = q.replace(f"/{conversation_command.value}", "").strip()
-
-    meta_log = (await ConversationAdapters.aget_conversation_by_user(user, request.user.client_app)).conversation_log
-
-    compiled_references, inferred_queries, defiltered_query = await extract_references_and_questions(
-        request, common, meta_log, q, (n or 5), (d or math.inf), conversation_command
-    )
-    online_results: Dict = dict()
-
-    if conversation_command == ConversationCommand.Default and is_none_or_empty(compiled_references):
-        conversation_command = ConversationCommand.General
-
-    elif conversation_command == ConversationCommand.Help:
-        conversation_config = await ConversationAdapters.aget_user_conversation_config(user)
-        if conversation_config == None:
-            conversation_config = await ConversationAdapters.aget_default_conversation_config()
-        model_type = conversation_config.model_type
-        formatted_help = help_message.format(model=model_type, version=state.khoj_version, device=get_device())
-        return StreamingResponse(iter([formatted_help]), media_type="text/event-stream", status_code=200)
-
-    elif conversation_command == ConversationCommand.Notes and not await EntryAdapters.auser_has_entries(user):
-        no_entries_found_format = no_entries_found.format()
-        if stream:
-            return StreamingResponse(iter([no_entries_found_format]), media_type="text/event-stream", status_code=200)
-        else:
-            response_obj = {"response": no_entries_found_format}
-            return Response(content=json.dumps(response_obj), media_type="text/plain", status_code=200)
-
-    elif conversation_command == ConversationCommand.Online:
-        try:
-            online_results = await search_with_google(defiltered_query, meta_log)
-        except ValueError as e:
-            return StreamingResponse(
-                iter(["Please set your SERPER_DEV_API_KEY to get started with online searches 🌐"]),
-                media_type="text/event-stream",
-                status_code=200,
-            )
-    elif conversation_command == ConversationCommand.Image:
-        update_telemetry_state(
-            request=request,
-            telemetry_type="api",
-            api="chat",
-            metadata={"conversation_command": conversation_command.value},
-            **common.__dict__,
-        )
-        image, status_code, improved_image_prompt = await text_to_image(q, meta_log)
-        if image is None:
-            content_obj = {
-                "image": image,
-                "intentType": "text-to-image",
-                "detail": "Failed to generate image. Make sure your image generation configuration is set.",
-            }
-            return Response(content=json.dumps(content_obj), media_type="application/json", status_code=status_code)
-        await sync_to_async(save_to_conversation_log)(
-            q,
-            image,
-            user,
-            meta_log,
-            intent_type="text-to-image",
-            inferred_queries=[improved_image_prompt],
-            client_application=request.user.client_app,
-        )
-        content_obj = {"image": image, "intentType": "text-to-image", "inferredQueries": [improved_image_prompt]}  # type: ignore
-        return Response(content=json.dumps(content_obj), media_type="application/json", status_code=status_code)
-
-    # Get the (streamed) chat response from the LLM of choice.
-    llm_response, chat_metadata = await agenerate_chat_response(
-        defiltered_query,
-        meta_log,
-        compiled_references,
-        online_results,
-        inferred_queries,
-        conversation_command,
-        user,
-        request.user.client_app,
-    )
-
-    chat_metadata.update({"conversation_command": conversation_command.value})
-
-    update_telemetry_state(
-        request=request,
-        telemetry_type="api",
-        api="chat",
-        metadata=chat_metadata,
-        **common.__dict__,
-    )
-
-    if llm_response is None:
-        return Response(content=llm_response, media_type="text/plain", status_code=500)
-
-    if stream:
-        return StreamingResponse(llm_response, media_type="text/event-stream", status_code=200)
-
-    iterator = AsyncIteratorWrapper(llm_response)
-
-    # Get the full response from the generator if the stream is not requested.
-    aggregated_gpt_response = ""
-    async for item in iterator:
-        if item is None:
-            break
-        aggregated_gpt_response += item
-
-    actual_response = aggregated_gpt_response.split("### compiled references:")[0]
-
-    response_obj = {"response": actual_response, "context": compiled_references}
-
-    return Response(content=json.dumps(response_obj), media_type="application/json", status_code=200)
 
 
 async def extract_references_and_questions(
