+++ src/khoj/routers/helpers.py
@@ -244,6 +244,7 @@ def generate_chat_response(
     conversation_command: ConversationCommand = ConversationCommand.Default,
     user: KhojUser = None,
     client_application: ClientApplication = None,
+    conversation_id: int = None,
 ) -> Tuple[Union[ThreadedGenerator, Iterator[str]], Dict[str, str]]:
     # Initialize Variables
     chat_response = None
@@ -261,6 +262,7 @@ def generate_chat_response(
             online_results=online_results,
             inferred_queries=inferred_queries,
             client_application=client_application,
+            conversation_id=conversation_id,
         )
 
         conversation_config = ConversationAdapters.get_valid_conversation_config(user)
